+++ src/DataProtection/test/New-AzDataProtectionBackupConfigurationClientObject.Tests.ps1
@@ -16,6 +16,16 @@ if(($null -eq $TestName) -or ($TestName -contains 'New-AzDataProtectionBackupCon
 
 Describe 'New-AzDataProtectionBackupConfigurationClientObject' {
     It '__AllParameterSets' -skip {
-        { throw [System.NotImplementedException] } | Should -Not -Throw
+        $subId = $env.TestBackupConfig.SubscriptionId
+        $storageAccountResourceGroup = $env.TestBackupConfig.StorageAccountResourceGroup
+        $storageAccountName = $env.TestBackupConfig.StorageAccountName
+
+        $storageAccount = Get-AzStorageAccount -ResourceGroupName $storageAccountResourceGroup -Name $storageAccountName
+        $containers = Get-AzStorageContainer -Context $storageAccount.Context
+        $backupConfig = New-AzDataProtectionBackupConfigurationClientObject -DatasourceType AzureBlob -StorageAccountName $storageAccountName -StorageAccountResourceGroupName $storageAccountResourceGroup -IncludeAllContainer
+
+        $backupConfig.ContainersList.Contains("abc") | Should be $true
+        $backupConfig.ContainersList.Contains("aab") | Should be $true
+        $backupConfig.ObjectType | Should be "BlobBackupDatasourceParameters"
     }
 }