+++ src/Resources/ResourceManager/Implementation/DeploymentStacks/SetAzManagmentGroupDeploymentStack.cs
@@ -101,12 +101,6 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
             {
                 Hashtable parameters = new Hashtable();
                 string filePath = "";
-                string parameterFilePath = "";
-
-                if (BicepUtility.IsBicepparamFile(TemplateParameterFile) && !BicepUtility.IsBicepFile(TemplateFile))
-                {
-                    throw new NotSupportedException($"Bicepparam file {TemplateParameterFile} is only supported with a Bicep template file");
-                }
 
                 switch (ParameterSetName)
                 {
@@ -123,13 +117,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                         break;
                     case ParameterFileTemplateSpecParameterSetName:
                     case ParameterFileTemplateUriParameterSetName:
-                        parameterFilePath = this.TryResolvePath(TemplateParameterFile);
-                        if (!File.Exists(parameterFilePath))
-                        {
-                            throw new PSInvalidOperationException(
-                                string.Format(ProjectResources.InvalidFilePath, TemplateParameterFile));
-                        }
-                        parameters = this.GetParameterObject(parameterFilePath);
+                        parameters = ResolveParameters();
 
                         // contruct the protected template URI if a query string was provided
                         if (!string.IsNullOrEmpty(QueryString))
@@ -141,6 +129,8 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                         }
                         break;
                     case ParameterFileTemplateFileParameterSetName:
+                        parameters = ResolveParameters();
+
                         filePath = this.TryResolvePath(TemplateFile);
                         if (!File.Exists(filePath))
                         {
@@ -149,14 +139,10 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                         }
                         filePath = ResolveBicepFile(filePath);
 
-                        parameterFilePath = this.TryResolvePath(TemplateParameterFile);
-                        if (!File.Exists(parameterFilePath))
-                        {
-                            throw new PSInvalidOperationException(
-                                string.Format(ProjectResources.InvalidFilePath, TemplateParameterFile));
-                        }
-                        parameters = this.GetParameterObject(ResolveBicepParameterFile(parameterFilePath));
                         TemplateUri = filePath;
+                        break;
+                    case ByParameterFileWithNoTemplateParameterSetName:
+                        parameters = ResolveParameters();
                         break;
                     case ParameterObjectTemplateFileParameterSetName:
                         filePath = this.TryResolvePath(TemplateFile);
@@ -214,6 +200,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                             location: Location,
                             templateUri: !string.IsNullOrEmpty(protectedTemplateUri) ? protectedTemplateUri : TemplateUri,
                             templateSpec: TemplateSpecId,
+                            templateJson: TemplateJson,
                             parameterUri: TemplateParameterUri,
                             parameters: parameters,
                             description: Description,