+++ src/Resources/Resources/ChangeLog.md
@@ -20,6 +20,10 @@
 
 ## Upcoming Release
 
+## Version 6.11.0
+* Supported `TemplateParameterFile` to accept a `.bicepparam` file.
+* Fixed inexplicable error message when subscription and scope are neither provided in `Get-AzRoleDefinition`. [#22716]
+
 ## Version 6.10.0
 * Added breaking change warnings for Azure Policy cmdlets.
 * Implemented logic that allows Deployment Stack objects to be piped into Save and Remove Deployment Stack cmdlets.